+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 * Add resolve claim to loan details screen. Refs UIU-1524.
 * Capitalize user status when displayed. Fixes UIU-1523.
 * Permissions -> Users: Create/reset password send. Refs UIU-1337.
+* Prevent change due date for claimed returned items. Refs UIU-1260.
 
 ## [3.0.0](https://github.com/folio-org/ui-users/tree/v3.0.0) (2020-03-17)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.26.0...v3.0.0)